requirements.txt
torchelastic==0.2.2
torchmetrics==0.6.0
absl-py==2.1.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
albumentations==1.3.0
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.8
anyio==4.2.0
appdirs==1.4.4
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
basicsr==1.4.2
beautifulsoup4==4.12.2
blinker==1.7.0
boltons==23.0.0
braceexpand==0.1.7
Brotli==1.0.9
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.4
click==8.1.7
contourpy==1.2.0
cryptography==41.0.3
cycler==0.12.1
decorator==5.1.1
dnspython==2.4.2
docker-pycreds==0.4.0
einops==0.3.0
exceptiongroup==1.0.4
executing==0.8.3
expecttest==0.1.6
fastapi==0.109.0
ffmpy==0.3.1
filelock==3.9.0
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.1.3
future==0.18.3
gitdb==4.0.11
GitPython==3.1.41
gmpy2==2.1.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
gradio==3.16.2
grpcio==1.60.0
h11==0.14.0
hf-xet==1.1.5
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.34.3
hypothesis==6.88.4
idna==3.4
imageio==2.33.1
imageio-ffmpeg==0.4.2
importlib-metadata==7.0.1
invisible-watermark==0.2.0
ipython==8.15.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.32
jsonpointer==2.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kornia==0.6.0
lazy_loader==0.3
libarchive-c==2.9
libmambapy==1.5.1
lightning-utilities==0.10.1
linkify-it-py==2.0.2
lmdb==1.4.1
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
more-itertools==8.12.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.26.0
oauthlib==3.2.2
omegaconf==2.1.1
open-clip-torch==2.0.2
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==3.4.18.65
orjson==3.9.12
packaging==23.1
pandas==2.2.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pip==23.3
pkginfo==1.9.6
pluggy==1.0.0
prettytable==3.6.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycosat==0.6.6
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
pyDeprecate==0.3.1
pydub==0.25.1
Pygments==2.15.1
Pympler==1.0.1
pyOpenSSL==23.2.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-etcd==0.4.5
python-multipart==0.0.6
pytorch-lightning==1.5.0
pytz==2023.3.post1
PyWavelets==1.5.0
PyYAML==6.0.1
qudida==0.0.4
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
safetensors==0.5.3
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.12.0
semver==3.0.2
sentry-sdk==1.39.2
setproctitle==1.3.3
setuptools==68.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.2.0
starlette==0.35.1
streamlit==1.12.1
streamlit-drawable-canvas==0.8.0
sympy==1.11.1
tb-nightly==2.16.0a20240126
tensorboard==2.15.1
tensorboard-data-server==0.7.2
test-tube==0.7.5
tf_keras-nightly==2.16.0.dev2024012610
threadpoolctl==3.2.0
tifffile==2023.12.9
timm==0.6.12
tokenizers==0.21.4
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.4
tqdm==4.65.0
traitlets==5.7.1
transformers==4.54.1
triton==2.1.0
truststore==0.8.0
types-dataclasses==0.6.6
typing_extensions==4.8.0
tzdata==2023.4
tzlocal==5.2
uc-micro-py==1.0.2
urllib3==1.26.18
uvicorn==0.27.0
validators==0.22.0
wandb==0.16.2
watchdog==3.0.0
wcwidth==0.2.13
webdataset==0.2.5
websockets==12.0
Werkzeug==3.0.1
wheel==0.41.2
yapf==0.32.0
yarl==1.9.4
zipp==3.17.0
zstandard==0.19.0
